Fix infinity ship drawing layer bug
Because the Game Engine does not support changing the draw order of
objects, all of the objects that are drawn to the screen must be
instantiated in the order they are drawn.

The Inifinity ship was creating new tiles as needed causing them to be
drawn over other objects on the screen. This commit changes the ship so
that it creates all the tiles to be drawn in the same order as the other
tiles causing their draw order to be stable throughout the game.