Remove rectangle inheirtance from GameWorldRegion
The game regions act as a primitive set rather than an individual
primitive. The calculations for the edges are similar for all game world
regions and can be placed into the base class.

Removed the multiple queues for the GameWorld. It's simpler to just
handle a single queue of game regions.