Change game world region initialization steps
Regions can now have their starting location set outside of the
constructor. Because the GameWorld is the only class that knows where a
newly created region needs to start, the region creation needed to be
performed inside of GameWorld.

With this change, regions can be instantiated in the user code and added
to the GameWorld to be initialized. This allows users to create new
regions and manually add them to the GameWorld without making changes to
GameWorld itself.